feat: complete parsers — typed events, checkpoints, rewind-snapshots, custom DB tables
- Add TypedEvent struct and parse_typed_events() with typed dispatch for all 24 event types
- Add extract_shutdown_data() (last shutdown) and extract_session_start() (first start)
- New checkpoints parser: parse index.md markdown table + load .md content bodies
- New rewind-snapshots parser: deserialize index.json with snapshot metadata
- Add generic read_custom_table() via PRAGMA table_info for arbitrary session.db tables
- Fix SessionStartData: version is Value (number), start_time is String (ISO 8601)
- Fix ToolExecCompleteData.error: can be object not just string
- Fix total_premium_requests/cost as f64 (fractional values observed in real data)
- Validated against all 53 real sessions: 0 parse failures across 126k+ events
- 20 unit tests covering all parsers, edge cases, and error handling

+++ crates/tracepilot-core/src/models/session_summary.rs
@@ -42,7 +42,7 @@ pub struct SessionSummary {
 #[serde(rename_all = "camelCase")]
 pub struct ShutdownMetrics {
     pub shutdown_type: Option<String>,
-    pub total_premium_requests: Option<u64>,
+    pub total_premium_requests: Option<f64>,
     pub total_api_duration_ms: Option<u64>,
     pub session_start_time: Option<u64>,
     pub current_model: Option<String>,
